+++ cli/src/controller/host-module.controller.ts
@@ -3,6 +3,7 @@ import { injectable } from 'inversify';
 import { ModuleReference, GahHost } from '@awdware/gah-shared';
 
 import { Controller } from './controller';
+import { ModuleReferenceHelper } from '../helper/module-reference-helper';
 
 @injectable()
 export class HostModuleController extends Controller {
@@ -17,38 +18,14 @@ export class HostModuleController extends Controller {
 
     this._loggerService.log('Adding new dependency to host');
 
-    const dependencyConfigPath_ = await this._promptService.input({
-      msg: 'Path to the gah-module.json of the new dependency',
-      enabled: () => !dependencyConfigPath,
-      cancelled: cancelled,
-      validator: (val: string) => {
-        if (!val.endsWith('gah-module.json')) { return false; }
-        return this._fileSystemService.fileExists(val);
-      }
-    });
+    const dependencyConfigPath_ = await ModuleReferenceHelper.askForGahModuleJson(this._promptService, this._fileSystemService, cancelled, dependencyConfigPath);
 
     cancelled = cancelled || !(dependencyConfigPath || dependencyConfigPath_);
     dependencyConfigPath = (dependencyConfigPath ?? dependencyConfigPath_).replace(/"/g, '');
 
     this._configService.readExternalConfig(dependencyConfigPath);
 
-    const availableExternalModules = this._configService.externalConfig.modules.map(x => x.name);
-    if (dependencyModuleNames && dependencyModuleNames.length > 0) {
-      const invalidModule = dependencyModuleNames.find(x => !availableExternalModules.includes(x));
-      if (invalidModule) {
-        throw new Error('Cannot find the module ' + invalidModule);
-      }
-    }
-    if (availableExternalModules.length === 1) {
-      dependencyModuleNames = availableExternalModules;
-    }
-
-    const dependencyModuleNames_ = await this._promptService.checkbox({
-      msg: 'Which of the modules do you want to add as a reference',
-      choices: () => availableExternalModules,
-      enabled: () => !dependencyModuleNames || dependencyModuleNames.length === 0,
-      cancelled: cancelled
-    });
+    const dependencyModuleNames_ = await ModuleReferenceHelper.askForModulesToAdd(this._configService, this._promptService, cancelled, dependencyModuleNames);
 
     if (!dependencyModuleNames || dependencyModuleNames.length === 0) { dependencyModuleNames = dependencyModuleNames_; }
